requirements.txt
beautifulsoup4
requests
argparse
python-dotenv
yfinance
streamlit
pandas